--- a/README.md
+++ b/README.md
@@ -57,7 +57,7 @@ Current support:
 |Type            | Number |
 |----------------|--------|
 |Licenses        | 184    |
-|Aliases         | 2550     |
+|Aliases         | 2549     |
 |Compatibilities | 23     |
 |Operators       | 14   |
 |Ambiguities     | 149 |