--- a/STATUS.md
+++ b/STATUS.md
@@ -2,7 +2,7 @@
 |Type            | Number |
 |----------------|--------|
 |Licenses        | 184    |
-|Aliases         | 2550     |
+|Aliases         | 2549     |
 |Compatibilities | 23     |
 |Operators       | 14   |
 |Ambiguities     | 149 |