--- a/var/licenses/Apache-2.0.json
+++ b/var/licenses/Apache-2.0.json
@@ -37,7 +37,6 @@
         "Apache Licence, version 2.0",
         "Apache-2.0 License",
         "Apache-2",
-        "Apache License  2.0",
         "Apache License v2.0",
         "Apache2",
         "Apache v2",