Pre-compile all regular expressions. This will speed things up significantly (after the first lookup). This will not impact the command line tool since it does only one lookup.


Testing with some lookups

## Before rewrite

```
real    0m47,035s
user    0m47,018s
sys     0m0,013s
```

## After rewrite

```
real    0m0,465s
user    0m0,452s
sys     0m0,013s
```